backend/requirements.txt
# Target: Python 3.11 (see Dockerfile). Used by: FastAPI API, Celery worker, Manim render, OCR/vision stack.
# Install: pip install -r requirements.txt

# --- Dev / test ---
pytest>=8.0
pytest-asyncio>=0.24

# --- HTTP API ---
cachetools>=5.3
fastapi>=0.115,<1
uvicorn[standard]>=0.30
python-multipart>=0.0.9
python-dotenv>=1.0
pydantic[email]>=2.4
email-validator>=2

# --- Auth / data / queue ---
openai>=1.40
supabase>=2.0
celery>=5.3
redis>=5
httpx>=0.27
websockets>=12

# --- Math & symbolic solver ---
sympy>=1.12
numpy>=1.26,<2
scipy>=1.11
opencv-python-headless>=4.8,<4.10

# --- Video (GeometryScene via CLI) ---
manim>=0.18,<0.20

# --- OCR & vision (orchestrator / legacy /ocr) ---
pix2tex>=0.1.4
paddleocr==2.7.3
paddlepaddle==2.6.2
ultralytics==8.2.2
